Separating views.py to files, kubernetes configuration
Moved milestones from views.py to milestone_views.py. Changed __init__.py for views. Added kubernetes configuration files.
