+++ kubernetes_config/mail-credentials.yaml
@@ -1,0 +1,10 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: mail-credentials
+type: Opaque
+data:
+  email_port: NTg3
+  email_host_user: 
+  email_host_password:
+  email_use_tls:
