+++ kubernetes_config/secret-key-credentials.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: secret-credentials
+type: Opaque
+data:
+  secret: dm4rKWolMGc1dnQrNCMheGk0Yl8pMip0cV45cjIpbSo2XnB0bSVAXyRpcnkqNWkqNSM
