+++ uks_app/static/uks_app/static.css
@@ -46,3 +46,4 @@ a:link.tab-link.active {
     border: 1px solid rgba(27,31,35,.15);
     box-shadow: none;
 }
+
