+++ uks_app/templates/uks_app/one_milestone.html
@@ -50,7 +50,7 @@
                 </div>
             {% endfor %}
         {% else %}
-            <p>There are no issues for this project.</p>
+            <p>There are no issues for this milestone.</p>
         {% endif %}
     </div>
 </div>
