+++ uks_app/views/__init__.py
@@ -2,9 +2,10 @@ from .project_views import *
 from .issue_views import *
 from .subissue_views import *
 from .label_views import *
-from .other_views import *
+from .views import *
 from .search_views import *
 from .code_change_views import *
 from .comment_views import *
 from .followers_views import *
 from .users_views import *
+from .milestone_views import *
